+++ pages/customer/CatalogCust.js
@@ -6,20 +6,6 @@ import database from '@react-native-firebase/database';
 
 const { width } = Dimensions.get('window');
 
-
-//product data from firebase
-// const items = [
-//   { id: 1, product_name:'Almond London', product_price: 5, product_image:''}
-//   // { id: 2, product_name: 'Suji', product_price: 6 },
-//   // { id: 3, product_name:'Tart Nenas', product_price: 5},
-//   // { id: 4, color: 'purple' },
-//   // { id: 5, color: 'orange' },
-//   // { id: 6, color: 'pink' },
-//   // { id: 7, color: 'brown' },
-//   // { id: 8, color: 'black' },
-//   // { id: 9, color: 'yellow' },
-// ];
-
 //define size for each block
 const ITEM_SIZE = 160;
 const ITEM_MARGIN = 10;
@@ -28,18 +14,20 @@ class CatalogCust extends Component {
 
 
   state = {
-    items:[ { id: 1, product_name:'Dummy', product_price: 5, product_image:'https://firebasestorage.googleapis.com/v0/b/whisk-n-bites-a4339.appspot.com/o/almondLondon.jpg?alt=media&token=dab20c87-17b0-4a6f-afe6-7049528402d7'}],
+    items: [],
   }
-  
-  getData=()=>{
+
+  getData = () => {
 
     database()
       .ref('/Products')
       .on('value', (snapshot) => {
         let data = [];
         snapshot.forEach((child) => {
-          data.push(child.val());
-          this.setItems(data);
+          temp = child.val()
+          temp.id = child.key
+          data.push(temp)
+          this.setItems(data)
           console.log(data)
         })
 
@@ -52,14 +40,13 @@ class CatalogCust extends Component {
     this.setState({ items: arr });
   }
 
-  // componentDidMount(){
-  //   this.getData()
-  // }
+  componentDidMount() {
+    this.getData()
+  }
 
-  
+
   renderItem = ({ id, product_name, product_price, product_image }) => (
-    // <View key={id} style={[styles.item, { backgroundColor: color }]} />
-    <CatalogBoxCust  key={id} product_name={product_name} product_price={product_price} product_image={product_image}/>
+    <CatalogBoxCust key={id} product_name={product_name} product_price={product_price} product_image={product_image} />
   );
 
   renderItems = () => {
@@ -85,9 +72,9 @@ class CatalogCust extends Component {
         <TabHeader title='Catalog' icon='cart' />
 
         <ScrollView>
-          {/* {((this.state.items)!==undefined)?<View style={styles.container}>{this.renderItems()}</View>:<View></View>} */}
+
           <View style={styles.container}>{this.renderItems()}</View>
-          
+
         </ScrollView>
 
       </SafeAreaView>